requirements.txt
bs4
requests
lxml
rich
PyCryptodomex
cx_Freeze